Move shapefile_readers.hpp files to legacy.
