+++ cpp/tests/CMakeLists.txt
@@ -115,9 +115,9 @@ set(POINT_IN_POLYGON_TEST_SRC
     "${CMAKE_CURRENT_SOURCE_DIR}/spatial/pip_compare.cu")
 ConfigureTest(POINT_IN_POLYGON_TEST "${POINT_IN_POLYGON_TEST_SRC}")
 
-set(SHAPEFILE_POYGON_READER_TEST_SRC
-    "${CMAKE_CURRENT_SOURCE_DIR}/io/read_shapefile_polygon_test.cu")
-ConfigureTest(SHAPEFILE_POLYGON_READER_TEST "${SHAPEFILE_POYGON_READER_TEST_SRC}")
+set(LEGACY_SHAPEFILE_POLYGON_READER_TEST_SRC
+    "${CMAKE_CURRENT_SOURCE_DIR}/io/legacy/read_shapefile_polygon_test.cu")
+ConfigureTest(LEGACY_SHAPEFILE_POLYGON_READER_TEST "${LEGACY_SHAPEFILE_POLYGON_READER_TEST_SRC}")
 
 set(SPATIAL_WINDOW_POINT_TEST_SRC
     "${CMAKE_CURRENT_SOURCE_DIR}/query/spatial_window_test_toy.cu")
