+++ cpp/tests/io/legacy/read_shapefile_polygon_test.cu
@@ -22,7 +22,7 @@
 #include <utilities/legacy/error_utils.hpp>
 #include <tests/utilities/legacy/cudf_test_utils.cuh>
 #include <tests/utilities/legacy/cudf_test_fixtures.h>
-#include <cuspatial/shapefile_readers.hpp>
+#include <cuspatial/legacy/shapefile_readers.hpp>
 #include <utility/utility.hpp>
